backend/requirements.txt
fastapi
python-decouple
paramiko